webrtc/protocol/unknown-mediatypes.html

<!doctype html>
<meta charset=utf-8>
<title>RTCPeerconnection SDP SCTP format test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../RTCPeerConnection-helper.js"></script>
<script>
'use strict';

promise_test(async t => {
  const caller = new RTCPeerConnection();
  const callee = new RTCPeerConnection();
  t.add_cleanup(() => caller.close());
  t.add_cleanup(() => callee.close());
  caller.createDataChannel('channel');
  const offer = await caller.createOffer();
  const [preamble, media_section, postamble] = offer.sdp.split('\r\nm=');
  assert_true(typeof(postamble) === 'undefined');
  assert_greater_than(media_section.search(
    /^application \d+ UDP\/DTLS\/SCTP webrtc-datachannel\r\n/), -1);
  assert_greater_than(media_section.search(/\r\na=sctp-port:\d+\r\n/), -1);
  assert_greater_than(media_section.search(/\r\na=mid:/), -1);
}, 'Generated Datachannel SDP uses correct SCTP offer syntax');

promise_test(async t => {
  const pc1 = new RTCPeerConnection();
  const pc2 = new RTCPeerConnection();
  t.add_cleanup(() => pc1.close());
  t.add_cleanup(() => pc2.close());
  pc1.addTransceiver('audio');
  pc1.createDataChannel('datachannel');
  const offer = await pc1.createOffer();
  await pc2.setRemoteDescription({
    type: 'offer',
    sdp: offer.sdp
        .replace('m=application ', 'm=unicorns ')
        .replace(/a=group:BUNDLE .*\r\n/, ''),
  });
  await pc1.setLocalDescription(offer);
  const answer = await pc2.createAnswer();
  await pc1.setRemoteDescription(answer);
  await pc2.setLocalDescription(answer);

  const [preamble, media_section, postamble] = offer.sdp.split('\r\nm=');
  assert_true(typeof(postamble) === 'undefined');
  assert_greater_than(media_section.search(
    /^unicorns 0/), -1);
}, 'Unknown media types are rejected with the port set to 0');
</script>
